refactor(backend): split vertex wire/policy and desktop gateway hop out of the ratcheted files
providers.py and desktop_proxy.py grew past the product line-count ratchet;
the vertex adapter's pure wire translation (vertex_wire.py), the PT policy
mixin (vertex_pt_policy.py), shared provider types (provider_types.py), and
the desktop BFF's gateway hop + Gemini body sanitization
(utils/llm/desktop_gemini_gateway.py) move into focused modules with the
original call sites and test seams preserved via re-exports.
