+++ backend/llm_gateway/gateway/providers.py
@@ -6,7 +6,6 @@ from collections.abc import Awaitable, Callable, Mapping
 from dataclasses import dataclass
 import json
 import os
-import re
 import time
 import logging
 from typing import Any, Protocol, cast
@@ -27,6 +26,28 @@ from llm_gateway.gateway.accounting import (
 )
 from llm_gateway.gateway.credentials import CredentialContext
 from llm_gateway.gateway.schemas import CredentialMode, FailureClass, ProviderRef, ProviderRejection
+from llm_gateway.gateway.provider_types import (  # noqa: F401 — re-exported provider contract
+    GENERIC_PROVIDER_FAILURE_MESSAGE,
+    ProviderFailure,
+    ProviderResponse,
+    _VertexHttpError,  # pyright: ignore[reportPrivateUsage]
+    _openai_usage_payload,  # pyright: ignore[reportPrivateUsage]
+)
+from llm_gateway.gateway.vertex_pt_policy import (  # noqa: F401 — re-exported vertex policy
+    VertexPTPolicyMixin,
+)
+from llm_gateway.gateway.vertex_wire import (  # noqa: F401 — re-exported wire contract
+    _nonnegative_int_or_zero,  # pyright: ignore[reportPrivateUsage]
+    _openai_sse_done,  # pyright: ignore[reportPrivateUsage, reportUnusedImport]
+    _text_content,  # pyright: ignore[reportPrivateUsage, reportUnusedImport]
+    _validate_embeddings_response_shape,  # pyright: ignore[reportPrivateUsage]
+    _vertex_embedding_predict_request,  # pyright: ignore[reportPrivateUsage]
+    _vertex_headers,  # pyright: ignore[reportPrivateUsage]
+    _vertex_predict_to_openai_embeddings,  # pyright: ignore[reportPrivateUsage]
+    _vertex_request,  # pyright: ignore[reportPrivateUsage]
+    _vertex_to_openai_response,  # pyright: ignore[reportPrivateUsage]
+    _vertex_to_openai_stream_chunk,  # pyright: ignore[reportPrivateUsage]
+)
 from llm_gateway.gateway.sse import SSEEventDecoder
 from utils.executors import critical_executor, run_blocking
 from utils.llm import vertex_pt_routing as ptr
@@ -41,12 +62,8 @@ DEFAULT_MAX_RESPONSE_BYTES = 2 * 1024 * 1024
 MAX_RESPONSE_BYTES_ENV_VAR = 'OPENAI_MAX_RESPONSE_BYTES'
 PROVIDER_ERROR_DETAIL_BYTES = 1000
 EXPOSE_PROVIDER_ERROR_DETAILS_ENV_VAR = 'LLM_GATEWAY_EXPOSE_PROVIDER_ERROR_DETAILS'
-GENERIC_PROVIDER_FAILURE_MESSAGE = 'provider request failed'
 GOOGLE_CLOUD_PROJECT_ENV_VAR = 'GOOGLE_CLOUD_PROJECT'
-GCP_LOCATION_ENV_VAR = 'GCP_LOCATION'
-DEFAULT_GCP_LOCATION = 'us-central1'
 GOOGLE_CLOUD_PLATFORM_SCOPE = 'https://www.googleapis.com/auth/cloud-platform'
-VERTEX_API_VERSION = 'v1'
 
 
 class ChatCompletionProvider(Protocol):
@@ -69,41 +86,6 @@ class EmbeddingProvider(Protocol):
         credentials: CredentialContext,
         timeout_ms: int,
     ) -> 'ProviderResponse': ...
-
-
-@dataclass
-class _VertexHttpError(Exception):
-    """A Vertex response the PT ladder may route around (429/404/5xx)."""
-
-    status_code: int
-    preview: bytes
-
-
-@dataclass
-class ProviderFailure(Exception):
-    failure_class: FailureClass
-    safe_message: str = GENERIC_PROVIDER_FAILURE_MESSAGE
-    provider_rejection: ProviderRejection = ProviderRejection.NONE
-
-    def __str__(self) -> str:
-        return self.safe_message
-
-
-@dataclass(frozen=True)
-class ProviderResponse(Mapping[str, Any]):
-    """OpenAI-compatible response plus provider-native accounting metadata."""
-
-    response: Mapping[str, Any]
-    accounting: ProviderResponseMetadata = ProviderResponseMetadata()
-
-    def __getitem__(self, key: str) -> Any:
-        return self.response[key]
-
-    def __iter__(self):
-        return iter(self.response)
-
-    def __len__(self) -> int:
-        return len(self.response)
 
 
 class OpenAICompatibleChatCompletionProvider:
@@ -246,7 +228,8 @@ class OpenAICompatibleChatCompletionProvider:
             raise ProviderFailure(FailureClass.PROVIDER_5XX_OMI_PAID) from exc
 
         _validate_embeddings_response_shape(parsed)
-        usage_raw = parsed.get('usage') if isinstance(parsed.get('usage'), Mapping) else {}
+        raw_usage = parsed.get('usage')
+        usage_raw = raw_usage if isinstance(raw_usage, Mapping) else {}
         prompt_tokens = _nonnegative_int_or_zero(usage_raw.get('prompt_tokens'))
         total_tokens = _nonnegative_int_or_zero(usage_raw.get('total_tokens'))
         return ProviderResponse(
@@ -311,7 +294,7 @@ class VertexAccessTokenSupplier:
         return token, expires_at
 
 
-class VertexGeminiProvider:
+class VertexGeminiProvider(VertexPTPolicyMixin):
     """Native Gemini-on-Vertex adapter behind the gateway's OpenAI contract.
 
     Also owns the company-paid desktop PT policy — pin, overflow ladder,
@@ -450,6 +433,7 @@ class VertexGeminiProvider:
         self._reject_byok(credentials)
         endpoint = self._endpoint(provider_ref.model, method='predict')
         payload = _vertex_embedding_predict_request(request)
+        parsed: Mapping[str, Any] | None = None
         try:
             headers = _vertex_headers(await self._vertex_access_token(), self._capacity_for(provider_ref.model))
             async with self._http_client.stream(
@@ -472,7 +456,7 @@ class VertexGeminiProvider:
         except httpx.HTTPError as exc:
             raise ProviderFailure(FailureClass.PROVIDER_5XX_OMI_PAID) from exc
 
-        normalized = _vertex_predict_to_openai_embeddings(parsed, model=provider_ref.model)
+        normalized = _vertex_predict_to_openai_embeddings(parsed or {}, model=provider_ref.model)
         _validate_embeddings_response_shape(normalized)
         # Vertex :predict reports billable characters, not tokens; the ledger
         # row records the request while usage stays NOT_REPORTED rather than
@@ -495,6 +479,7 @@ class VertexGeminiProvider:
         deadline = self._now() + max(timeout_ms, 0) / 1000.0
         attempts = self._attempt_plan(anchor)
         last_error: _VertexHttpError | None = None
+        parsed: Mapping[str, Any] | None = None
         while attempts:
             model, capacity = attempts.pop(0)
             remaining_ms = int((deadline - self._now()) * 1000)
@@ -517,6 +502,7 @@ class VertexGeminiProvider:
                     continue
                 _raise_for_status(error.status_code, error.preview, credential_mode=credentials.mode)
             self._record_model_available(model)
+            assert parsed is not None
             return parsed
         assert last_error is not None
         _raise_for_status(last_error.status_code, last_error.preview, credential_mode=credentials.mode)
@@ -556,155 +542,6 @@ class VertexGeminiProvider:
         except httpx.HTTPError as exc:
             raise ProviderFailure(FailureClass.PROVIDER_5XX_OMI_PAID) from exc
 
-    # --- PT policy (all decisions delegate to vertex_pt_routing) -----------
-
-    def _attempt_plan(self, anchor: str) -> list[tuple[str, str]]:
-        serving = self._serving_model(anchor)
-        return [(serving, self._capacity_for(serving))]
-
-    def _serving_model(self, anchor: str) -> str:
-        intended = ptr.desktop_serving_model(
-            anchor,
-            target_dedicated_ready=self._pt_target_is_ready(),
-            override=self._env(self._pt_model_override_env),
-        )
-        return self._first_reachable(intended)
-
-    def _provisioned_model(self) -> str:
-        return ptr.resolve_pt_model(
-            target_dedicated_ready=self._pt_target_is_ready(),
-            override=self._env(self._pt_model_override_env),
-        )
-
-    def _capacity_for(self, model: str) -> str:
-        return ptr.request_type_for(model=model, pt_model=self._provisioned_model())
-
-    def _recovery_attempts(self, served_model: str, status_code: int, preview: bytes) -> list[tuple[str, str]]:
-        message = _bounded_error_text(preview)
-        if ptr.is_model_unavailable(status_code, message):
-            self._record_model_unavailable(served_model)
-            return [(rung, ptr.REQUEST_TYPE_SHARED) for rung in self._fallback_chain(served_model)]
-        if self._overflow_triggered(status_code, message):
-            return self._overflow_plan(served_model)
-        return []
-
-    def _observe_attempt(self, model: str, capacity: str, status_code: int, preview: bytes) -> None:
-        """Latch PT-target probe outcomes from a dedicated attempt."""
-        if capacity != ptr.REQUEST_TYPE_DEDICATED:
-            return
-        message = _bounded_error_text(preview)
-        unavailable = ptr.is_model_unavailable(status_code, message)
-        exhausted = self._overflow_triggered(status_code, message)
-        if not unavailable:
-            self._record_pt_target_observation(not exhausted)
-
-    def _overflow_triggered(self, status_code: int, message: str) -> bool:
-        return ptr.is_provisioned_capacity_exhausted(status_code, message) or ptr.is_provisioned_capacity_absent(
-            status_code, message
-        )
-
-    def _overflow_plan(self, served_model: str) -> list[tuple[str, str]]:
-        if not self._overflow_enabled():
-            return []
-        pt_model = self._provisioned_model()
-        if served_model != pt_model:
-            return []
-        try:
-            ladder = ptr.resolve_overflow_ladder(
-                pt_model=pt_model, override=self._env(self._overflow_model_override_env)
-            )
-        except ValueError:
-            return []
-        plan: list[tuple[str, str]] = []
-        for rung in ladder:
-            if not self._model_believed_available(rung):
-                continue
-            if rung == ptr.PT_MODEL_TARGET and self._pt_probe_due():
-                plan.append((rung, ptr.REQUEST_TYPE_DEDICATED))
-            plan.append((rung, ptr.REQUEST_TYPE_SHARED))
-        return plan
-
-    def _fallback_chain(self, model: str) -> tuple[str, ...]:
-        try:
-            return ptr.resolve_fallback_chain(
-                model=model,
-                pt_model=self._provisioned_model(),
-                unreachable=self._unreachable_models(),
-                override=self._env(self._overflow_model_override_env),
-            )
-        except ValueError:
-            return ()
-
-    def _pt_target_is_ready(self) -> bool:
-        return self._pt_target_ready and self._model_believed_available(ptr.PT_MODEL_TARGET)
-
-    def _model_believed_available(self, model: str) -> bool:
-        observed = self._model_unavailable_at.get(model)
-        if observed is None:
-            return True
-        return (self._now() - observed) >= self._probe_ttl_seconds
-
-    def _unreachable_models(self) -> frozenset[str]:
-        return frozenset(model for model in self._model_unavailable_at if not self._model_believed_available(model))
-
-    def _first_reachable(self, model: str) -> str:
-        if self._model_believed_available(model):
-            return model
-        for rung in self._fallback_chain(model):
-            if self._model_believed_available(rung):
-                return rung
-        return model
-
-    def _record_model_unavailable(self, model: str) -> None:
-        self._model_unavailable_at[model] = self._now()
-        if model == ptr.PT_MODEL_TARGET:
-            self._pt_target_ready = False
-
-    def _record_model_available(self, model: str) -> None:
-        self._model_unavailable_at.pop(model, None)
-
-    def _pt_probe_due(self) -> bool:
-        if self._pt_target_probed_at is None:
-            return True
-        return (self._now() - self._pt_target_probed_at) >= self._probe_ttl_seconds
-
-    def _record_pt_target_observation(self, ready: bool) -> None:
-        became_ready = ready and not self._pt_target_ready
-        self._pt_target_ready = ready
-        self._pt_target_probed_at = self._now()
-        if became_ready:
-            logger.info('llm_gateway vertex pt_promotion target=%s', ptr.PT_MODEL_TARGET)
-
-    def _overflow_enabled(self) -> bool:
-        return self._env(self._overflow_enabled_env, 'true').strip().lower() not in {'0', 'false', 'no', 'off'}
-
-    def _multi_region_location(self) -> str:
-        return (
-            self._env(self._multi_region_location_env, ptr.MULTI_REGION_LOCATION).strip() or ptr.MULTI_REGION_LOCATION
-        )
-
-    @staticmethod
-    def _env(name: str, default: str = '') -> str:
-        return os.getenv(name, default)
-
-    def _endpoint(self, model: str, *, method: str) -> str:
-        project = os.getenv(self._project_env, '').strip()
-        if not project:
-            raise ProviderFailure(FailureClass.INVALID_CONFIG)
-        # Gemini 3.x has no regional endpoint: it needs the un-prefixed host
-        # plus a multi-region `locations/{loc}` path segment. Building a
-        # regional URL for it is what made every 3.x request 404 in
-        # production on 2026-08-18 (see vertex_pt_routing).
-        host, location = ptr.vertex_endpoint(
-            model=model,
-            regional_location=os.getenv(self._location_env, DEFAULT_GCP_LOCATION).strip() or DEFAULT_GCP_LOCATION,
-            multi_region_location=self._multi_region_location(),
-        )
-        return (
-            f'https://{host}/{VERTEX_API_VERSION}/projects/{project}'
-            f'/locations/{location}/publishers/google/models/{model}:{method}'
-        )
-
     async def _vertex_access_token(self) -> str:
         try:
             return await self._access_token_supplier()
@@ -717,416 +554,6 @@ class VertexGeminiProvider:
     def _reject_byok(credentials: CredentialContext) -> None:
         if credentials.mode == CredentialMode.BYOK:
             raise ProviderFailure(FailureClass.BYOK_UNSUPPORTED_PROVIDER)
-
-
-def _vertex_headers(access_token: str, capacity: str) -> dict[str, str]:
-    if not access_token.strip():
-        raise ProviderFailure(FailureClass.INVALID_CONFIG)
-    # Without the capacity header Vertex silently spills over-cap dedicated
-    # requests onto pay-as-you-go; asking for `dedicated` turns that into a
-    # 429 the PT ladder can act on, and everything else is pinned `shared` so
-    # it can never draw down the reservation.
-    return {
-        'Authorization': f'Bearer {access_token}',
-        'Content-Type': 'application/json',
-        ptr.REQUEST_TYPE_HEADER: capacity,
-    }
-
-
-def _vertex_request(request: Mapping[str, Any]) -> dict[str, Any]:
-    unsupported_params = sorted(
-        key
-        for key in (
-            'frequency_penalty',
-            'logit_bias',
-            'logprobs',
-            'n',
-            'presence_penalty',
-            'prompt_cache_key',
-            'seed',
-            'top_logprobs',
-            'user',
-        )
-        if key in request
-    )
-    if unsupported_params:
-        raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-
-    system_parts: list[dict[str, str]] = []
-    contents: list[dict[str, Any]] = []
-    raw_messages = request.get('messages')
-    if not isinstance(raw_messages, list):
-        raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-    tool_names_by_id: dict[str, str] = {}
-    for message in raw_messages:
-        if not isinstance(message, Mapping):
-            raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-        role = message.get('role')
-        if role == 'system':
-            # Vertex systemInstruction takes text parts only. _vertex_parts raises rather
-            # than silently flattening an image here, same as everywhere else below.
-            system_parts.extend(_system_text_parts(message.get('content')))
-            continue
-        if role == 'tool':
-            contents.append(_vertex_function_response_content(message, tool_names_by_id))
-            continue
-        if role == 'assistant' and isinstance(message.get('tool_calls'), list):
-            content, names = _vertex_model_tool_call_content(message, tool_names_by_id)
-            tool_names_by_id.update(names)
-            contents.append(content)
-            continue
-        if role not in {'user', 'assistant'}:
-            raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-        contents.append(
-            {
-                'role': 'model' if role == 'assistant' else 'user',
-                'parts': _vertex_parts(message.get('content')),
-            }
-        )
-
-    generation_config: dict[str, Any] = {}
-    for request_key, vertex_key in (('temperature', 'temperature'), ('top_p', 'topP')):
-        if request_key in request:
-            generation_config[vertex_key] = request[request_key]
-    if 'stop' in request:
-        stop = request['stop']
-        if isinstance(stop, str):
-            generation_config['stopSequences'] = [stop]
-        elif isinstance(stop, list) and all(isinstance(item, str) for item in stop):
-            generation_config['stopSequences'] = stop
-        else:
-            raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-    output_limit = _output_limit(request)
-    if output_limit is not None:
-        generation_config['maxOutputTokens'] = output_limit
-    thinking_budget = _thinking_budget(request)
-    if thinking_budget is not None:
-        generation_config['thinkingConfig'] = {'thinkingBudget': thinking_budget}
-    response_format = request.get('response_format')
-    if isinstance(response_format, Mapping):
-        format_type = response_format.get('type')
-        if format_type == 'json_object':
-            generation_config['responseMimeType'] = 'application/json'
-        else:
-            json_schema = response_format.get('json_schema')
-            if not isinstance(json_schema, Mapping) or not isinstance(json_schema.get('schema'), Mapping):
-                raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-            generation_config['responseMimeType'] = 'application/json'
-            generation_config['responseSchema'] = dict(cast(Mapping[str, Any], json_schema['schema']))
-
-    payload: dict[str, Any] = {'contents': contents}
-    if system_parts:
-        payload['systemInstruction'] = {'parts': system_parts}
-    if generation_config:
-        payload['generationConfig'] = generation_config
-    tools = _vertex_tools(request.get('tools'))
-    if tools is not None:
-        payload['tools'] = tools
-    tool_config = _vertex_tool_config(request.get('tool_choice'))
-    if tool_config is not None:
-        payload['toolConfig'] = tool_config
-    return payload
-
-
-def _output_limit(request: Mapping[str, Any]) -> int | None:
-    max_completion_tokens = request.get('max_completion_tokens')
-    max_tokens = request.get('max_tokens')
-    value = max_completion_tokens if max_completion_tokens is not None else max_tokens
-    if value is None:
-        return None
-    if not isinstance(value, int) or isinstance(value, bool) or value <= 0:
-        raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-    return value
-
-
-def _thinking_budget(request: Mapping[str, Any]) -> int | None:
-    if request.get('reasoning_effort') == 'none':
-        return 0
-    # The OpenAI SDK's extra_body convention flattens `extra_body={'google': …}`
-    # into a top-level `google` field, so per-request Gemini options arrive both
-    # ways: as a forwarded top-level `google` param and via provider_options.
-    google_options = request.get('google')
-    if isinstance(google_options, Mapping):
-        budget = _thinking_budget_from_google(google_options)
-        if budget is not None:
-            return budget
-    extra_body = request.get('extra_body')
-    if isinstance(extra_body, Mapping):
-        extra_google = extra_body.get('google')
-        if isinstance(extra_google, Mapping):
-            budget = _thinking_budget_from_google(extra_google)
-            if budget is not None:
-                return budget
-    return None
-
-
-def _thinking_budget_from_google(google_options: Mapping[str, Any]) -> int | None:
-    thinking_config = google_options.get('thinking_config')
-    if not isinstance(thinking_config, Mapping):
-        return None
-    thinking_budget = thinking_config.get('thinking_budget')
-    if not isinstance(thinking_budget, int) or isinstance(thinking_budget, bool) or thinking_budget < 0:
-        raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-    return thinking_budget
-
-
-def _vertex_tools(value: Any) -> list[dict[str, Any]] | None:
-    """Translate OpenAI function tools into a Gemini tools declaration."""
-    if value is None:
-        return None
-    if not isinstance(value, list) or not value:
-        raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-    declarations: list[dict[str, Any]] = []
-    for tool in value:
-        if not isinstance(tool, Mapping) or tool.get('type') != 'function':
-            raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-        function = tool.get('function')
-        if not isinstance(function, Mapping) or not isinstance(function.get('name'), str) or not function['name']:
-            raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-        declaration: dict[str, Any] = {'name': function['name']}
-        description = function.get('description')
-        if isinstance(description, str) and description:
-            declaration['description'] = description
-        parameters = function.get('parameters')
-        if isinstance(parameters, Mapping) and parameters:
-            declaration['parameters'] = dict(cast(Mapping[str, Any], parameters))
-        declarations.append(declaration)
-    return [{'functionDeclarations': declarations}]
-
-
-def _vertex_tool_config(value: Any) -> dict[str, Any] | None:
-    """Translate OpenAI tool_choice into Gemini functionCallingConfig."""
-    if value is None:
-        return None
-    if value == 'required':
-        return {'functionCallingConfig': {'mode': 'ANY'}}
-    if value == 'auto':
-        return {'functionCallingConfig': {'mode': 'AUTO'}}
-    if value == 'none':
-        return {'functionCallingConfig': {'mode': 'NONE'}}
-    if isinstance(value, Mapping) and value.get('type') == 'function':
-        function = value.get('function')
-        if isinstance(function, Mapping) and isinstance(function.get('name'), str) and function['name']:
-            return {'functionCallingConfig': {'mode': 'ANY', 'allowedFunctionNames': [function['name']]}}
-    raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-
-
-def _vertex_model_tool_call_content(
-    message: Mapping[str, Any],
-    tool_names_by_id: dict[str, str],
-) -> tuple[dict[str, Any], dict[str, str]]:
-    """An assistant message with OpenAI tool_calls -> a Gemini model functionCall content."""
-    parts: list[dict[str, Any]] = []
-    if isinstance(message.get('content'), str) and message['content']:
-        parts.append({'text': message['content']})
-    names: dict[str, str] = {}
-    for call in message['tool_calls']:
-        if not isinstance(call, Mapping) or call.get('type') != 'function':
-            raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-        function = call.get('function')
-        if not isinstance(function, Mapping) or not isinstance(function.get('name'), str) or not function['name']:
-            raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-        raw_arguments = function.get('arguments')
-        if isinstance(raw_arguments, Mapping):
-            arguments: dict[str, Any] = dict(cast(Mapping[str, Any], raw_arguments))
-        elif isinstance(raw_arguments, str) and raw_arguments:
-            try:
-                decoded = json.loads(raw_arguments)
-            except json.JSONDecodeError as exc:
-                raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH) from exc
-            if not isinstance(decoded, Mapping):
-                raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-            arguments = dict(cast(Mapping[str, Any], decoded))
-        else:
-            arguments = {}
-        parts.append({'functionCall': {'name': function['name'], 'args': arguments}})
-        call_id = call.get('id')
-        if isinstance(call_id, str) and call_id:
-            names[call_id] = function['name']
-    if not parts:
-        parts = [{'text': ''}]
-    return {'role': 'model', 'parts': parts}, names
-
-
-def _vertex_function_response_content(
-    message: Mapping[str, Any],
-    tool_names_by_id: dict[str, str],
-) -> dict[str, Any]:
-    """An OpenAI tool-result message -> a Gemini user functionResponse content."""
-    raw_content = message.get('content')
-    if isinstance(raw_content, Mapping):
-        response_payload: dict[str, Any] = dict(cast(Mapping[str, Any], raw_content))
-    elif isinstance(raw_content, str) and raw_content:
-        try:
-            decoded = json.loads(raw_content)
-        except json.JSONDecodeError:
-            response_payload = {'result': raw_content}
-        else:
-            response_payload = (
-                dict(cast(Mapping[str, Any], decoded)) if isinstance(decoded, Mapping) else {'result': raw_content}
-            )
-    else:
-        response_payload = {}
-    name = message.get('name')
-    if not isinstance(name, str) or not name:
-        name = tool_names_by_id.get(str(message.get('tool_call_id') or ''), '')
-    if not name:
-        raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-
-    return {'role': 'user', 'parts': [{'functionResponse': {'name': name, 'response': response_payload}}]}
-
-
-def _nonnegative_int_or_zero(value: object) -> int:
-    return value if isinstance(value, int) and not isinstance(value, bool) and value > 0 else 0
-
-
-def _bounded_error_text(preview: bytes) -> str:
-    return preview.decode('utf-8', errors='replace')
-
-
-def _vertex_embedding_predict_request(request: Mapping[str, Any]) -> dict[str, Any]:
-    """An OpenAI embeddings request -> a Vertex :predict instances payload."""
-    inputs = request.get('input')
-    if isinstance(inputs, str):
-        inputs = [inputs]
-    if not isinstance(inputs, list) or not inputs:
-        raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-    instances: list[dict[str, Any]] = []
-    for text in inputs:
-        if not isinstance(text, str) or not text:
-            raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-        instance: dict[str, Any] = {'content': text}
-        task_type = request.get('task_type')
-        if isinstance(task_type, str) and task_type:
-            instance['task_type'] = task_type
-        title = request.get('title')
-        if isinstance(title, str) and title:
-            instance['title'] = title
-        instances.append(instance)
-    return {'instances': instances}
-
-
-def _vertex_predict_to_openai_embeddings(response: Mapping[str, Any], *, model: str) -> dict[str, Any]:
-    predictions = response.get('predictions')
-    data: list[dict[str, Any]] = []
-    if isinstance(predictions, list):
-        for index, prediction in enumerate(predictions):
-            embeddings = prediction.get('embeddings') if isinstance(prediction, Mapping) else None
-            values = embeddings.get('values') if isinstance(embeddings, Mapping) else None
-            if not isinstance(values, list):
-                values = []
-            data.append({'object': 'embedding', 'embedding': [float(value) for value in values], 'index': index})
-    return {'object': 'list', 'data': data, 'model': model}
-
-
-def _validate_embeddings_response_shape(response: Mapping[str, Any]) -> None:
-    if response.get('object') != 'list' or not isinstance(response.get('data'), list) or not response['data']:
-        raise ProviderFailure(FailureClass.PROVIDER_5XX_OMI_PAID)
-    for item in response['data']:
-        if not isinstance(item, Mapping) or not isinstance(item.get('embedding'), list):
-            raise ProviderFailure(FailureClass.PROVIDER_5XX_OMI_PAID)
-
-
-def _vertex_to_openai_response(
-    response: Mapping[str, Any],
-    *,
-    requested_model: str,
-    usage: ProviderUsage | None = None,
-) -> dict[str, Any]:
-    candidates = response.get('candidates')
-    candidate = (
-        candidates[0] if isinstance(candidates, list) and candidates and isinstance(candidates[0], Mapping) else None
-    )
-    content = _vertex_candidate_text(candidate)
-    finish_reason = _vertex_finish_reason(candidate.get('finishReason') if candidate is not None else 'SAFETY')
-    normalized: dict[str, Any] = {
-        'id': str(response.get('responseId') or 'vertex_gateway'),
-        'object': 'chat.completion',
-        'created': int(time.time()),
-        'model': requested_model,
-        'choices': [
-            {
-                'index': 0,
-                'message': {'role': 'assistant', 'content': content},
-                'finish_reason': finish_reason,
-            }
-        ],
-    }
-    if usage is not None:
-        normalized['usage'] = _openai_usage_payload(usage)
-    return normalized
-
-
-def _vertex_to_openai_stream_chunk(
-    response: Mapping[str, Any],
-    *,
-    requested_model: str,
-    usage: ProviderUsage | None = None,
-) -> tuple[bytes | None, bool]:
-    candidates = response.get('candidates')
-    candidate = (
-        candidates[0] if isinstance(candidates, list) and candidates and isinstance(candidates[0], Mapping) else None
-    )
-    if candidate is None and usage is None:
-        return None, False
-    text = _vertex_candidate_text(candidate)
-    raw_finish_reason = candidate.get('finishReason') if candidate is not None else None
-    finish_reason = _vertex_finish_reason(raw_finish_reason) if raw_finish_reason else None
-    if not text and finish_reason is None and usage is None:
-        return None, False
-    body: dict[str, Any] = {
-        'id': str(response.get('responseId') or 'vertex_gateway'),
-        'object': 'chat.completion.chunk',
-        'created': int(time.time()),
-        'model': requested_model,
-        'choices': (
-            [
-                {
-                    'index': 0,
-                    'delta': {'content': text} if text else {},
-                    'finish_reason': finish_reason,
-                }
-            ]
-            if candidate is not None
-            else []
-        ),
-    }
-    if usage is not None:
-        body['usage'] = _openai_usage_payload(usage)
-    return _openai_sse(body), finish_reason is not None
-
-
-def _vertex_candidate_text(candidate: Mapping[str, Any] | None) -> str:
-    if candidate is None:
-        return ''
-    content = candidate.get('content')
-    if not isinstance(content, Mapping):
-        return ''
-    parts = content.get('parts')
-    if not isinstance(parts, list):
-        return ''
-    text_parts: list[str] = []
-    for part in parts:
-        if isinstance(part, Mapping) and isinstance(part.get('text'), str):
-            text_parts.append(part['text'])
-    return ''.join(text_parts)
-
-
-def _vertex_finish_reason(value: object) -> str:
-    normalized = str(value or '').upper()
-    if normalized in {'MAX_TOKENS', 'LENGTH'}:
-        return 'length'
-    if normalized in {'SAFETY', 'BLOCKLIST', 'PROHIBITED_CONTENT', 'SPII', 'RECITATION'}:
-        return 'content_filter'
-    return 'stop'
-
-
-def _openai_sse(body: Mapping[str, Any]) -> bytes:
-    return f'data: {json.dumps(dict(body), separators=(",", ":"))}\n\n'.encode('utf-8')
-
-
-def _openai_sse_done() -> bytes:
-    return b'data: [DONE]\n\n'
 
 
 class AnthropicMessagesProvider:
@@ -1280,107 +707,6 @@ def _anthropic_to_openai_response(
     if usage is not None:
         normalized['usage'] = _openai_usage_payload(usage)
     return normalized
-
-
-def _openai_usage_payload(usage: ProviderUsage) -> dict[str, Any]:
-    return {
-        'prompt_tokens': usage.prompt_tokens,
-        'completion_tokens': usage.output_tokens + usage.reasoning_tokens,
-        'total_tokens': usage.total_tokens,
-        'prompt_tokens_details': {'cached_tokens': usage.cached_input_tokens},
-        'completion_tokens_details': {'reasoning_tokens': usage.reasoning_tokens},
-    }
-
-
-# RFC 2397 permits parameters between the media type and the base64 token
-# (`data:image/jpeg;charset=utf-8;base64,...`), and browser- or canvas-produced
-# data URLs do emit them. Rejecting those would be the mirror of the bug this
-# module just fixed: refusing an image we can in fact represent.
-_VERTEX_DATA_URL_RE = re.compile(
-    r'^data:(?P<mime>[\w.+-]+/[\w.+-]+)(?:;[\w.+-]+=[^;,]*)*;(?i:base64),(?P<data>.+)$',
-    re.DOTALL,
-)
-
-
-def _vertex_parts(content: Any) -> list[dict[str, Any]]:
-    """Translate OpenAI-shaped message content into Vertex parts.
-
-    Anything this cannot represent raises CAPABILITY_MISMATCH rather than being
-    dropped. That distinction is the whole point of this function: the previous
-    implementation ran every message through _text_content(), which keeps only
-    `type == "text"` parts, so an image attached to a Gemini request vanished
-    silently and the model answered about content it never received. For a
-    caller like utils/screen_frames/judge.py — a privacy gate that decides
-    whether a screenshot may be stored — a confident answer from a model that
-    was sent no image is worse than an error, because the caller's fail-closed
-    handling never triggers.
-    """
-    if content is None:
-        # See the empty-parts note at the end of this function: None is what an
-        # assistant tool-call turn carries, and Vertex rejects a Content with no parts.
-        return [{'text': ''}]
-    if isinstance(content, str):
-        return [{'text': content}]
-    if not isinstance(content, list):
-        raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-
-    parts: list[dict[str, Any]] = []
-    for part in cast(list[object], content):
-        if not isinstance(part, Mapping):
-            raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-        typed_part = cast(Mapping[str, Any], part)
-        part_type = typed_part.get('type')
-        if part_type == 'text':
-            text = typed_part.get('text')
-            if not isinstance(text, str):
-                raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-            parts.append({'text': text})
-            continue
-        if part_type == 'image_url':
-            image_url = typed_part.get('image_url')
-            if not isinstance(image_url, Mapping):
-                raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-            url = cast(Mapping[str, Any], image_url).get('url')
-            if not isinstance(url, str):
-                raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-            match = _VERTEX_DATA_URL_RE.match(url)
-            if match is None:
-                # A remote https:// image is not fetchable by Vertex the way it is by
-                # OpenAI; only inline bytes and gs:// URIs are. Refuse rather than send
-                # a request the model will answer without the image.
-                raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-            parts.append({'inlineData': {'mimeType': match.group('mime'), 'data': match.group('data')}})
-            continue
-        raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-    # A message with no representable content still needs one part: Vertex rejects a
-    # Content with an empty parts array, and the previous implementation always
-    # produced [{'text': ''}] here (via _text_content(None) == ''). An assistant
-    # tool-call turn carries content=None, so this path is reachable the moment a
-    # multi-turn Gemini feature exists.
-    return parts or [{'text': ''}]
-
-
-def _system_text_parts(content: Any) -> list[dict[str, str]]:
-    parts = _vertex_parts(content)
-    for part in parts:
-        if 'text' not in part:
-            raise ProviderFailure(FailureClass.CAPABILITY_MISMATCH)
-    return [{'text': cast(str, part['text'])} for part in parts] or [{'text': ''}]
-
-
-def _text_content(content: Any) -> str:
-    if isinstance(content, str):
-        return content
-    if isinstance(content, list):
-        parts: list[str] = []
-        for part in cast(list[object], content):
-            if not isinstance(part, Mapping):
-                continue
-            typed_part = cast(Mapping[str, Any], part)
-            if typed_part.get('type') == 'text' and isinstance(typed_part.get('text'), str):
-                parts.append(typed_part['text'])
-        return '\n'.join(parts)
-    return ''
 
 
 def _openai_finish_reason(stop_reason: Any) -> str:
